requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
llama-cpp-python==0.2.20
faiss-cpu==1.10.0
sentence-transformers>=2.2.2
notion-client==2.0.0
atlassian-python-api>=3.41.0
pydantic<3.0.0,>=2.0.0
lxml>=4.9.0
tqdm>=4.66.1
huggingface_hub>=0.19.0
hf_transfer>=0.1.5
langdetect>=1.0.9
requests>=2.31.0

# 모델 다운로드
# https://huggingface.co/meta-llama/Meta-Llama-3-8B-Instruct